+++ Cargo.toml
@@ -8,17 +8,18 @@ edition = "2018"
 default = ["service-actix"]
 ui = []
 profile-index = ["flame", "flamer"]
-service-actix = ["actix-files", "actix-http", "actix-rt", "actix-web"]
+service-actix = ["actix-files", "actix-http", "actix-web"]
 service-rocket = ["rocket", "rocket_contrib"]
 
 [dependencies]
 actix-files = { version = "0.2", optional = true }
 actix-http = { version = "1.0", optional = true }
 actix-web = { version = "2.0", optional = true }
-actix-rt = { version = "1.0", optional = true }
+actix-rt = { version = "1.0"}
 anyhow = "1.0"
 ape = "0.2.0"
 app_dirs = "1.1.1"
+async-trait = "0.1.22"
 base64 = "0.11.0"
 diesel = { version = "1.4", features = ["sqlite", "r2d2"] }
 diesel_migrations = { version = "1.4", features = ["sqlite"] }
