+++ src/service/mod.rs
@@ -2,6 +2,9 @@ mod constants;
 mod dto;
 mod error;
 
+#[cfg(test)]
+mod tests;
+
 #[cfg(feature = "service-actix")]
 mod actix;
 #[cfg(feature = "service-actix")]
