+++ src/service/rocket/mod.rs
@@ -6,8 +6,4 @@ pub mod server;
 #[cfg(test)]
 mod api_tests;
 #[cfg(test)]
-mod swagger;
-#[cfg(test)]
-mod test;
-#[cfg(test)]
-mod web;
+pub mod test;
